requirements.txt
attrs==24.2.0
cattrs==23.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
confluent-kafka==2.5.0
exceptiongroup==1.2.2
fiona==1.9.6
flatbuffers==24.3.25
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
idna==3.7
joblib==1.4.2
networkx==3.3
numpy==1.26.4
openmeteo_requests==1.3.0
openmeteo_sdk==1.14.1
osmnx==1.9.4
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
py4j==0.10.9.7
pyproj==3.6.1
pyspark==3.5.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
scikit-learn==1.5.1
scipy==1.14.1
shapely==2.0.6
simplejson==3.19.3
six==1.16.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.2
